Redis Validations

+++ src/autoGenerate/graphql/controllers/RedisController.js
@@ -2,15 +2,27 @@ import redis from '../../../redis';
 import MasterController from './MasterController';
 
 class RedisController extends MasterController {
+  REDIS_SUCCESS_STATE = ['ready', 'connect'];
+
   constructor(authentication) {
     const model = '';
     super(model, authentication);
     this.redis = redis;
   }
 
+  validateRedisConn() {
+    if (this.redis && this.REDIS_SUCCESS_STATE.includes(this.redis.status)) {
+      return true;
+    } 
+    return false;
+  }
+
   async get(hkey) {
-    let data = await this.redis.get(hkey);
-    return JSON.parse(data);
+    if (this.validateRedisConn()) {
+      let data = await this.redis.get(hkey);
+      return JSON.parse(data);
+    }
+    return null;
   }
  
   async set(obj, { hkey , maxAge } = {}) {
@@ -23,11 +35,11 @@ class RedisController extends MasterController {
   }
 
   async destroy(hkey) {
-    await this.redis.del(hkey);
+    if (this.validateRedisConn()) await this.redis.del(hkey);
   }
 
   async flush() {
-    await redis.flushall();
+    if (this.validateRedisConn()) await redis.flushall();
   }
 }
